# openapi.yaml
openapi: 3.1.0
info:
  title: shortlink
  version: 1.0.0
  description: |
    Short links over HTTP: redirects with expiry and click limits, click statistics, QR codes as
    SVG or PNG.

    ## Auth zones
    This service has two genuinely separate auth zones:

    - **Management API** (`/v1/*`, `GET /metrics`) — Bearer API key (`Authorization: Bearer <secret>`).
      Keys carry a role: `read`, `write`, or `readwrite` (satisfies both `read`- and `write`-gated
      routes). A key without the required role gets `403 FORBIDDEN`; a missing/invalid key gets
      `401 UNAUTHORIZED`.
    - **Public zone** (`GET /{code}`, `GET /{code}/qr`, `GET /robots.txt`) — no authentication at
      all, reachable by any browser, rate limited per client address instead of per key. `GET
      /health`, `GET /ready` and `GET /v1/info` are also unauthenticated (operational/discovery
      endpoints), but sit outside both the `/v1` management surface and the public redirect/QR
      surface.

    ## Error envelope
    Every error response, from every route, uses one shape: `{ "error": { "code", "message",
    "details"? } }`. `details` is present only for a handful of codes (validation failures, link
    gone with its status, patch failure context) — omitted otherwise.
...
paths:
...
  /v1/links:
    post:
      operationId: shortlink.links.create
      summary: Create a short link
      description: |
        Requires the `write` role. `slug` is optional — omitted, a random unbiased base62 code of
        the configured length is generated (retried up to 5 times against collisions/reserved
        words before failing hard with a 500, which is not a modeled `LinkError`). When given, the
        slug must match `^[A-Za-z0-9][A-Za-z0-9_-]{2,63}$` (identical to the request-body schema's
        own pattern for `slug`, so a malformed slug is rejected as `400 VALIDATION_FAILED` before
        the domain layer ever runs — the domain layer's own "bad format" `SLUG_RESERVED` check
        exists only as defense-in-depth for callers of the domain service outside this HTTP API,
        and is not reachable through this endpoint). A syntactically valid slug can still fail as
        `SLUG_RESERVED` when it collides with a reserved platform word (`v1`, `health`, `ready`,
        `metrics`, `qr`, `api`, `admin`, `favicon.ico`, `robots.txt`, `sitemap.xml`, `.well-known`,
        case-insensitively) or as `SLUG_TAKEN` when already in use.
      tags: [Links]
      security:
        - ApiKeyAuth: []
      x-stack-visibility: authenticated
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkCreate'
      responses:
        '201':
          description: Link created.
          headers:
            Location:
              description: '`/v1/links/{code}` of the new link.'
              schema: { type: string }
          content:
            application/json:
              schema:
                type: object
                required: [link]
                properties:
                  link:
                    $ref: '#/components/schemas/Link'
        '400':
          description: Bad request — one of `VALIDATION_FAILED`, `INVALID_URL`, `HOST_BLOCKED`, `INVALID_EXPIRY`, or `SLUG_RESERVED` (see description).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/SlugTaken'
        '429':
          $ref: '#/components/responses/RateLimited'

    get:
      operationId: shortlink.links.list
      summary: List links
      description: Requires `read` access. Keyset pagination, newest first.
      tags: [Links]
      security:
        - ApiKeyAuth: []
      x-stack-visibility: authenticated
      parameters:
        - $ref: '#/components/parameters/ListQ'
        - $ref: '#/components/parameters/ListTag'
        - $ref: '#/components/parameters/ListStatus'
        - $ref: '#/components/parameters/ListCreatedBy'
        - $ref: '#/components/parameters/ListLimit'
        - $ref: '#/components/parameters/ListCursor'
      responses:
        '200':
          description: Page of links.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkList'
        '400':
          description: '`VALIDATION_FAILED` (bad query parameter) or `INVALID_CURSOR` (malformed `cursor`).'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/links/{code}:
    get:
      operationId: shortlink.links.get
      summary: Get a link
      description: Requires `read` access.
      tags: [Links]
      security:
        - ApiKeyAuth: []
      x-stack-visibility: authenticated
      parameters:
        - $ref: '#/components/parameters/CodeParam'
      responses:
        '200':
          description: The link.
          content:
            application/json:
              schema:
                type: object
                required: [link]
                properties:
                  link:
                    $ref: '#/components/schemas/Link'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/LinkNotFound'
        '429':
          $ref: '#/components/responses/RateLimited'

    patch:
      operationId: shortlink.links.patch
      summary: Update a link
      description: |
        Requires the `write` role. Partial update — every property is optional but at least one
        must be present; only the properties given are changed. `enabled` and `slug`-independent
        fields only; the code/slug itself is never patchable.
      tags: [Links]
      security:
        - ApiKeyAuth: []
      x-stack-visibility: authenticated
      parameters:
        - $ref: '#/components/parameters/CodeParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkPatch'
      responses:
        '200':
          description: The updated link.
          content:
            application/json:
              schema:
                type: object
                required: [link]
                properties:
                  link:
                    $ref: '#/components/schemas/Link'
        '400':
          description: One of `VALIDATION_FAILED`, `INVALID_URL`, or `INVALID_EXPIRY`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/LinkNotFound'
        '429':
          $ref: '#/components/responses/RateLimited'

    delete:
      operationId: shortlink.links.delete
      summary: Delete a link
      description: Requires the `write` role. Permanent — the code becomes available for reuse.
      tags: [Links]
      security:
        - ApiKeyAuth: []
      x-stack-visibility: authenticated
      parameters:
        - $ref: '#/components/parameters/CodeParam'
      responses:
        '204':
          description: Deleted. No body.
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/LinkNotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/links/{code}/stats:
    get:
      operationId: shortlink.links.getStats
      summary: Click statistics for one link
      description: Requires `read` access. `days` is the lookback window (default 7 is service-wide; this route's own default is 30).
      tags: [Stats]
      security:
        - ApiKeyAuth: []
      x-stack-visibility: authenticated
      parameters:
        - $ref: '#/components/parameters/CodeParam'
        - $ref: '#/components/parameters/DaysParam'
      responses:
        '200':
          description: Click statistics.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkStats'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/LinkNotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/links/{code}/qr:
    get:
      operationId: shortlink.links.getQr
      summary: QR code for a link (management)
      description: |
        Requires `read` access. Unlike the public `GET /{code}/qr`, this works on a link in **any**
        status (active, disabled, expired, exhausted) — it is a management convenience, not a
        gate on whether the link still redirects. Encodes the link's short URL, not arbitrary text.
      tags: [QR]
      security:
        - ApiKeyAuth: []
      x-stack-visibility: authenticated
      parameters:
        - $ref: '#/components/parameters/CodeParam'
        - $ref: '#/components/parameters/QrFormat'
        - $ref: '#/components/parameters/QrScale'
        - $ref: '#/components/parameters/QrMargin'
      responses:
        '200':
          $ref: '#/components/responses/QrImage'
        '304':
          $ref: '#/components/responses/QrNotModified'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/LinkNotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/qr:
    get:
      operationId: shortlink.qr.getForText
      summary: QR code for arbitrary text
      description: |
        Requires `read` access. Encodes the caller-supplied `text` query parameter verbatim — not
        tied to any link/code. `413`-equivalent case: text that does not fit the largest supported
        QR version (213 bytes, UTF-8 encoded) fails as `400 QR_TOO_LONG`.
      tags: [QR]
      security:
        - ApiKeyAuth: []
      x-stack-visibility: authenticated
      parameters:
        - $ref: '#/components/parameters/QrText'
        - $ref: '#/components/parameters/QrFormat'
        - $ref: '#/components/parameters/QrScale'
        - $ref: '#/components/parameters/QrMargin'
      responses:
        '200':
          $ref: '#/components/responses/QrImage'
        '304':
          $ref: '#/components/responses/QrNotModified'
        '400':
          description: '`VALIDATION_FAILED` (missing/malformed query, including a missing `text`) or `QR_TOO_LONG` (text exceeds QR capacity).'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/stats:
    get:
      operationId: shortlink.stats.get
      summary: Service-wide click statistics
      description: Requires `read` access. `days` lookback window, default 7.
      tags: [Stats]
      security:
        - ApiKeyAuth: []
      x-stack-visibility: authenticated
      parameters:
        - $ref: '#/components/parameters/DaysParam'
      responses:
        '200':
          description: Overview statistics.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsOverview'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'

  # ---------------------------------------------------------------------------------------------
  # Public zone — no authentication, rate limited per client address (separate bucket from /v1).
  # ---------------------------------------------------------------------------------------------
  /{code}/qr:
    get:
      operationId: shortlink.qr.getPublic
      summary: QR code for a link (public)
      description: |
        No authentication. Works only when the link's current status is `active` — any other
        status (disabled, expired, exhausted) responds `410 LINK_GONE` instead of an image, unlike
        the management `GET /v1/links/{code}/qr`, which works on any status. An unknown or
        malformed code responds `404 LINK_NOT_FOUND`. Rate limited per client address (the public
        bucket, `REDIRECT_RATE_LIMIT_MAX`, distinct from the `/v1` per-key bucket).
      tags: [QR]
      security: []
      x-stack-visibility: public
      parameters:
        - $ref: '#/components/parameters/PublicCodeParam'
        - $ref: '#/components/parameters/QrFormat'
        - $ref: '#/components/parameters/QrScale'
        - $ref: '#/components/parameters/QrMargin'
      responses:
        '200':
          $ref: '#/components/responses/QrImage'
        '304':
          $ref: '#/components/responses/QrNotModified'
        '404':
          $ref: '#/components/responses/LinkNotFound'
        '410':
          $ref: '#/components/responses/LinkGone'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: |
        `Authorization: Bearer <secret>`. Every configured key is compared in constant time.
        Missing or unrecognized secret: `401 UNAUTHORIZED` with a `WWW-Authenticate: Bearer`
        response header. Recognized key lacking the role a route requires: `403 FORBIDDEN`.
  parameters:
    CodeParam:
      name: code
      in: path
      required: true
      description: The link's short code.
      schema:
        $ref: '#/components/schemas/Code'
...
    QrFormat:
      name: format
      in: query
      required: false
      schema:
        type: string
        enum: [svg, png]
        default: svg
...
    DaysParam:
      name: days
      in: query
      required: false
      description: Lookback window in days, 1-399.
      schema:
        type: string
        pattern: '^([1-9]|[1-9][0-9]|[1-3][0-9][0-9])$'
    ListQ:
      name: q
      in: query
      required: false
      description: Free-text search over code and URL.
      schema: { type: string, minLength: 1, maxLength: 200 }
    ListTag:
      name: tag
      in: query
      required: false
      schema: { type: string, minLength: 1, maxLength: 40 }
    ListStatus:
      name: status
      in: query
      required: false
      schema:
        type: string
        enum: [active, disabled, expired, exhausted]
    ListCreatedBy:
      name: createdBy
      in: query
      required: false
      description: Filter by the API key id that created the link.
      schema:
        type: string
        pattern: '^[A-Za-z0-9_-]{1,64}$'
    ListLimit:
      name: limit
      in: query
      required: false
      description: Page size, 1-200. Default 50.
      schema:
        type: string
        pattern: '^([1-9]|[1-9][0-9]|1[0-9][0-9]|200)$'
    ListCursor:
      name: cursor
      in: query
      required: false
      description: Opaque keyset cursor from a previous page's `nextCursor`.
      schema: { type: string, maxLength: 200 }
  schemas:
...
    Link:
      type: object
      description: The canonical link read shape, returned by every route that echoes a link.
      required:
        - code
        - shortUrl
        - previewUrl
        - qrUrl
        - url
        - permanent
        - enabled
        - status
        - expiresAt
        - maxClicks
        - clicks
        - remainingClicks
        - lastClickAt
        - tags
        - note
        - createdBy
        - createdAt
        - updatedAt
      properties:
        code:
          $ref: '#/components/schemas/Code'
        shortUrl: { type: string, description: 'Public redirect URL (`{publicBaseUrl}/{code}`).' }
        previewUrl: { type: string, description: 'Public preview URL (`{publicBaseUrl}/{code}+`).' }
        qrUrl: { type: string, description: 'Public QR URL (`{publicBaseUrl}/{code}/qr`).' }
        url: { type: string, description: Destination URL. }
        permanent: { type: boolean, description: Whether the public redirect uses 301 (true) or 302 (false). }
        enabled: { type: boolean }
        status:
          $ref: '#/components/schemas/LinkStatus'
        expiresAt: { type: [string, 'null'], format: date-time }
        maxClicks: { type: [integer, 'null'] }
        clicks: { type: integer, description: Total recorded clicks (bots included). }
        remainingClicks: { type: [integer, 'null'], description: '`maxClicks - clicks`, floored at 0; `null` when `maxClicks` is unset.' }
        lastClickAt: { type: [string, 'null'], format: date-time }
        tags:
          type: array
          items: { type: string }
          description: Deduplicated, lower-cased, sorted.
        note: { type: [string, 'null'] }
        createdBy: { type: string, description: Id of the API key that created the link. }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    LinkCreate:
      type: object
      additionalProperties: false
      required: [url]
      properties:
        url: { type: string, minLength: 8, maxLength: 8192, description: Must be an absolute http/https URL with no embedded credentials, not pointing at this service's own host, and not on the configured block list. }
        slug:
          $ref: '#/components/schemas/Code'
        permanent: { type: boolean, description: 'Defaults to the service''s `DEFAULT_PERMANENT` config when omitted.' }
        expiresAt: { type: [string, 'null'], maxLength: 40, description: ISO 8601 timestamp, must be in the future. }
        maxClicks: { type: [integer, 'null'], minimum: 1, maximum: 1000000000 }
        tags:
          type: array
          maxItems: 20
          items: { type: string, minLength: 1, maxLength: 40, pattern: '^[^,]+$' }
        note: { type: [string, 'null'], maxLength: 500 }
    LinkPatch:
      type: object
      additionalProperties: false
      minProperties: 1
      description: Every property is optional; at least one must be present. `code`/`slug` is never patchable.
      properties:
        url: { type: string, minLength: 8, maxLength: 8192 }
        permanent: { type: boolean }
        enabled: { type: boolean }
        expiresAt: { type: [string, 'null'], maxLength: 40 }
        maxClicks: { type: [integer, 'null'], minimum: 1, maximum: 1000000000 }
        tags:
          type: array
          maxItems: 20
          items: { type: string, minLength: 1, maxLength: 40, pattern: '^[^,]+$' }
        note: { type: [string, 'null'], maxLength: 500 }
    LinkList:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Link'
        nextCursor:
          type: [string, 'null']
          description: Opaque keyset cursor for the next page, or `null` when this is the last page.
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, description: 'Stable machine-readable code, e.g. `LINK_NOT_FOUND`.' }
            message: { type: string }
            details:
              description: Present only for a subset of codes (validation failures, link-gone status, patch context).
  responses:
    Unauthorized:
      description: Missing or unrecognized API key.
      headers:
        WWW-Authenticate:
          schema: { type: string, example: Bearer }
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error: { code: UNAUTHORIZED, message: missing or invalid API key }
    Forbidden:
      description: Key is valid but lacks the role this route requires.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error: { code: FORBIDDEN, message: 'this API key has no write access' }
    ValidationFailed:
      description: Request failed JSON-Schema validation.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorEnvelope'
            example:
              error:
                code: VALIDATION_FAILED
                message: body/url must NOT have fewer than 8 characters
                details:
                  - path: /url
                    message: must NOT have fewer than 8 characters
                    params: { limit: 8 }
    RateLimited:
      description: Self-throttle exceeded (per API key on `/v1/*`, per client address on the public zone).
      headers:
        X-RateLimit-Limit:
          schema: { type: integer }
        X-RateLimit-Remaining:
          schema: { type: integer }
        X-RateLimit-Reset:
          schema: { type: integer }
        Retry-After:
          schema: { type: integer }
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error: { code: RATE_LIMITED, message: 'rate limit exceeded, retry in 42000ms' }
    LinkNotFound:
      description: No link with this code exists.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error: { code: LINK_NOT_FOUND, message: link not found }
...
    SlugTaken:
      description: Requested `slug` is already in use by another link.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error: { code: SLUG_TAKEN, message: '"launch" is already in use' }
...